backend/requirements.txt
aiofiles==0.5.0
alembic==1.6.2
amqp==5.0.6
aniso8601==7.0.0
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
autoflake==1.4
bcrypt==3.2.0
billiard==3.6.4.0
black==21.5b1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.65.1
flake8==3.9.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
httptools==0.1.2
idna==2.10
invoke==1.5.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
Mako==1.1.4
MarkupSafe==2.0.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
orjson==3.5.2
passlib==1.7.4
pathspec==0.8.1
promise==2.3
prompt-toolkit==3.0.18
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.15
sqlalchemy-mixins==1.3
starlette==0.14.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
